requirements.txt
# InvokeAI Cleanup Dependencies
streamlit
pyyaml
